+++ yash-semantics/src/expansion/initial/arith.rs
@@ -19,22 +19,55 @@
 use super::super::attr::AttrChar;
 use super::super::attr::Origin;
 use super::super::phrase::Phrase;
+use super::super::ErrorCause;
 use super::Env;
 use super::Error;
 use crate::expansion::expand_text;
+use std::rc::Rc;
 use yash_arith::eval;
+use yash_env::variable::ReadOnlyError;
+use yash_env::variable::Scope::Global;
+use yash_env::variable::Value::Scalar;
+use yash_env::variable::Variable;
+use yash_env::variable::VariableSet;
+use yash_syntax::source::Code;
 use yash_syntax::source::Location;
+use yash_syntax::source::Source;
 use yash_syntax::syntax::Text;
 
-pub async fn expand(text: &Text, _location: &Location, env: &mut Env<'_>) -> Result<Phrase, Error> {
+struct VarEnv<'a>(&'a mut VariableSet);
+
+impl<'a> yash_arith::Env for VarEnv<'a> {
+    type AssignVariableError = ReadOnlyError;
+
+    #[rustfmt::skip]
+    fn get_variable(&self, name: &str) -> Option<&str> {
+        if let Some(Variable { value: Scalar(value), .. }) = self.0.get(name) {
+            Some(value)
+        } else {
+            None
+        }
+    }
+
+    fn assign_variable(&mut self, name: &str, value: String) -> Result<(), ReadOnlyError> {
+        let value = Variable {
+            value: Scalar(value),
+            last_assigned_location: None, // TODO Provide correct location
+            is_exported: false,
+            read_only_location: None,
+        };
+        self.0.assign(Global, name.to_owned(), value).map(drop)
+    }
+}
+
+pub async fn expand(text: &Text, location: &Location, env: &mut Env<'_>) -> Result<Phrase, Error> {
     let (expression, exit_status) = expand_text(env.inner, text).await?;
     if exit_status.is_some() {
         env.last_command_subst_exit_status = exit_status;
     }
 
-    let result = eval(&expression);
+    let result = eval(&expression, &mut VarEnv(&mut env.inner.variables));
 
-    // TODO Test this
     match result {
         Ok(value) => {
             let value = value.to_string();
@@ -49,18 +82,30 @@ pub async fn expand(text: &Text, _location: &Location, env: &mut Env<'_>) -> Res
                 .collect();
             Ok(Phrase::Field(chars))
         }
-        Err(error) => todo!("handle error: {}", error),
+        Err(error) => Err(Error {
+            cause: ErrorCause::ArithError(error.cause),
+            location: Location {
+                code: Rc::new(Code {
+                    value: expression.into(),
+                    start_line_number: 1.try_into().unwrap(),
+                    source: Source::Arith {
+                        original: location.clone(),
+                    },
+                }),
+                range: error.location,
+            },
+        }),
     }
 }
 
 #[cfg(test)]
 mod tests {
-    use super::super::super::ErrorCause;
     use super::*;
     use crate::tests::echo_builtin;
     use crate::tests::in_virtual_system;
     use crate::tests::return_builtin;
     use futures_util::FutureExt;
+    use yash_arith::TokenError;
     use yash_env::semantics::ExitStatus;
     use yash_env::system::Errno;
 
@@ -123,5 +168,22 @@ mod tests {
         assert_eq!(e.cause, ErrorCause::CommandSubstError(Errno::ENOSYS));
         assert_eq!(*e.location.code.value.borrow(), "$(x)");
         assert_eq!(e.location.range, 0..4);
+    }
+
+    #[test]
+    fn error_in_arithmetic_evaluation() {
+        let text = "09".parse().unwrap();
+        let location = Location::dummy("my location");
+        let mut env = yash_env::Env::new_virtual();
+        let mut env = Env::new(&mut env);
+        let result = expand(&text, &location, &mut env).now_or_never().unwrap();
+        let e = result.unwrap_err();
+        assert_eq!(
+            e.cause,
+            ErrorCause::ArithError(TokenError::InvalidNumericConstant.into())
+        );
+        assert_eq!(*e.location.code.value.borrow(), "09");
+        assert_eq!(e.location.code.source, Source::Arith { original: location });
+        assert_eq!(e.location.range, 0..2);
     }
 }